move get_ui_url to _shared_files.utils
