+++ covalent_ui/result_webhook.py
@@ -26,20 +26,12 @@ import requests
 import covalent_ui.app as ui_server
 from covalent._results_manager import Result
 from covalent._shared_files import logger
-from covalent._shared_files.config import get_config
+from covalent._shared_files.utils import get_ui_url
 from covalent_dispatcher._db.dispatchdb import encode_dict, extract_graph, extract_metadata
-
-DEFAULT_UI_ADDRESS = get_config("user_interface.address")
-DEFAULT_UI_PORT = get_config("user_interface.port")
 
 app_log = logger.app_log
 
 # UI server webhook for result updates
-
-
-def get_ui_url(path):
-    baseUrl = f"http://{DEFAULT_UI_ADDRESS}:{DEFAULT_UI_PORT}"
-    return f"{baseUrl}{path}"
 
 
 async def send_update(result: Result) -> None:
